+++ saffier/core/db/__init__.py
@@ -1,1 +1,6 @@
+from .context_vars import set_schema as set_schema
 from .context_vars import set_tenant as set_tenant
+from .context_vars import with_tenant as with_tenant
+from .querysets.mixins import with_schema as with_schema
+
+__all__ = ["set_schema", "set_tenant", "with_schema", "with_tenant"]